added edit staff, fixed header and body layout, configured delete user permission

+++ users/views.py
@@ -361,12 +361,15 @@ def edit_user_profile(request, user_id):
 
     # user needs to be deleted
     if request.method == 'POST' and 'delete_user_id' in request.POST:
-        user_id_to_delete = request.POST['delete_user_id']
-        user_to_delete = User.objects.get(id=user_id_to_delete)
-        user_to_delete.delete()
-        messages.success(request, f"User {
-            user_to_delete.username} has been deleted.")
-        return redirect('list')
+        if not request.user.is_superuser and request.user.role != 'admin':
+            messages.error(request, "You do not have permission to delete.")
+        else:
+            user_id_to_delete = request.POST['delete_user_id']
+            user_to_delete = User.objects.get(id=user_id_to_delete)
+            user_to_delete.delete()
+            messages.success(request, f"User {
+                user_to_delete.username} has been deleted.")
+            return redirect('list')
 
     if request.method == 'POST':
         user_form = UserEditForm(
